Fix tile system and debug tools
Fixed a bug in the tile system in which the mouse centering and debug rectangles were offset by 0.5*tile_width along the y-axis. With the fix the Mouse is now centered on tile in x and y, while Entities remain centered on tile bottom edge in x. This makes the debug tools for rectangles, tiles and coordinates usable again.